chore: move suite wallet types and constants to common packages
