+++ packages/suite/src/actions/suite/storageActions.ts
@@ -6,7 +6,7 @@ import * as suiteActions from '@suite-actions/suiteActions';
 import { serializeDiscovery, serializeDevice } from '@suite-utils/storage';
 import { deviceGraphDataFilterFn } from '@wallet-utils/graphUtils';
 import { getFormDraftKey } from '@wallet-utils/formDraftUtils';
-import { FormDraftPrefixKeyValues } from '@wallet-constants/formDraft';
+import { FormDraftPrefixKeyValues } from '@suite-common/wallet-constants';
 import { STORAGE } from './constants';
 
 import type { Dispatch, GetState, AppState, TrezorDevice } from '@suite-types';
