--- packages/suite/src/constants/wallet/notification.ts
@@ -1,8 +1,0 @@
-export default {
-    PRIORITY: {
-        error: 2,
-        warning: 1,
-        info: 2,
-        success: 3,
-    },
-};
