+++ packages/suite/src/types/wallet/backend.ts
@@ -1,18 +1,1 @@
-import { NETWORKS } from '@wallet-config';
-
-/**
- * @deprecated
- */
-export type BlockbookUrl = {
-    coin: string;
-    url: string;
-    tor?: boolean; // Added by TOR
-};
-
-export type BackendType = 'blockbook' | 'electrum' | 'ripple' | 'blockfrost';
-
-export type CustomBackend = {
-    coin: typeof NETWORKS[number]['symbol'];
-    type: BackendType;
-    urls: string[];
-};
+export type { BlockbookUrl, BackendType, CustomBackend } from '@suite-common/wallet-types';
