+++ packages/suite/src/types/wallet/polling.ts
@@ -1,11 +1,1 @@
-import type { PollingPrefixKeyValues } from '@wallet-constants/polling';
-
-export type PollingKeyPrefix = typeof PollingPrefixKeyValues[number];
-export type PollingKey = `${PollingKeyPrefix}/${string}`;
-export type Polling = {
-    pollingFunction: () => void;
-    intervalMs: number;
-    timeoutId?: number;
-    counter: number;
-    maxPollingRequestCount: number;
-};
+export type { PollingKeyPrefix, PollingKey, Polling } from '@suite-common/wallet-types';
