+++ suite-common/wallet-constants/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@suite-common/wallet-constants",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint": "eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    }
+}
