+++ suite-common/wallet-constants/src/polling.ts
