+++ suite-common/wallet-types/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@suite-common/wallet-types",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint": "eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    },
+    "dependencies": {
+        "@suite-common/wallet-networks-config": "*",
+        "@suite-common/wallet-constants": "*",
+        "@suite-common/intl-types": "*",
+        "@trezor/connect": "9.0.0-beta.4",
+        "react": "17.0.2",
+        "react-hook-form": "6.15.7"
+    }
+}
